Initial commit, assg03 update with GitHub workflow
Initial commit of assignment 03.  Should be basically
ready to be used as an assignment template, barring
typos or bugs.

Update the workflow a bit from previous assignment.
Now trying out using unittest functions
for testing/workflow.  Result is maybe a little cleaner,
can put all tasks in 1 file and tests in another.
But on the other hand, may be harder for a student
to understand when a test fails from a similar
doctest.