Add in basic assignment description
Add in a basic readme / assignment description
for assignment 03.  Can probably continue to improve this,
but for these class assignments, probably most
important to have objectives and a generic description.
Most of the actual assignment description for students
will remain in the ipython notebook that they will
put their work into.